Escape context according to CQ specs. Closes #26
