+++ src/main/kotlin/tech/mihoyo/mirai/data/common/MessageDTO.kt
@@ -19,6 +19,7 @@ import net.mamoe.mirai.message.*
 import net.mamoe.mirai.message.data.*
 import net.mamoe.mirai.utils.MiraiExperimentalAPI
 import net.mamoe.mirai.utils.currentTimeMillis
+import tech.mihoyo.mirai.util.toCQString
 import java.net.URL
 
 /*
@@ -61,29 +62,6 @@ data class CQPrivateMessagePacketDTO(
     override var post_type: String = "message"
     val message_type: String = "private"
 }
-
-
-// Message toCQString
-@MiraiExperimentalAPI
-suspend fun Message.toCQString(): String {
-    return when (this) {
-        is PlainText -> content
-        is At -> "[CQ:at,qq=$target]"
-        is Face -> "[CQ:face,id=$id]"
-        is VipFace -> "[CQ:vipface,id=${kind.id},name=${kind.name},count=${count}]"
-        is PokeMessage -> "[CQ:poke,id=${id},type=${type},name=${name}]"
-        is AtAll -> "[CQ:at,qq=all]"
-        is Image -> {
-            "[CQ:image,file=$imageId,url=${queryUrl()}]"
-        }
-        is RichMessage -> "[CQ:rich,data=${content}]"
-        is MessageSource -> ""
-        is QuoteReply -> ""
-        is Voice -> "[CQ:voice,url=${url},md5=${md5},file=${fileName}]"
-        else -> "此处消息的转义尚未被插件支持"
-    }
-}
-
 
 // Message DTO
 @Serializable
@@ -233,6 +211,7 @@ sealed class MessageDTO : DTO
 /*
     Extend function
  */
+@MiraiExperimentalAPI
 suspend fun MessageEvent.toDTO(isRawMessage: Boolean = false): CQEventDTO {
     val rawMessage = WrappedCQMessageChainString("")
     message.forEach { rawMessage.value += it.toCQString() }
