+++ src/main/kotlin/tech/mihoyo/mirai/util/CQMessgeParser.kt
@@ -34,6 +34,7 @@ import net.mamoe.mirai.contact.Contact
 import net.mamoe.mirai.getGroupOrNull
 import net.mamoe.mirai.message.data.*
 import net.mamoe.mirai.message.uploadImage
+import net.mamoe.mirai.utils.MiraiExperimentalAPI
 import java.io.ByteArrayInputStream
 import java.io.File
 import java.net.URL
@@ -288,6 +289,24 @@ private fun String.toMap(): HashMap<String, String> {
         map[parts[0]] = parts[1].unescape()
     }
     return map
+}
+
+@MiraiExperimentalAPI
+suspend fun Message.toCQString(): String {
+    return when (this) {
+        is PlainText -> content.escape()
+        is At -> "[CQ:at,qq=$target]"
+        is Face -> "[CQ:face,id=$id]"
+        is VipFace -> "[CQ:vipface,id=${kind.id},name=${kind.name},count=${count}]"
+        is PokeMessage -> "[CQ:poke,id=${id},type=${type},name=${name}]"
+        is AtAll -> "[CQ:at,qq=all]"
+        is Image -> "[CQ:image,file=$imageId,url=${queryUrl()}]"  // 无需转义 https://github.com/richardchien/coolq-http-api/blob/30bfedec692cd1c383ad7561af80c4744343861f/src/cqhttp/plugins/message_enhancer/message_enhancer.cpp#L249
+        is RichMessage -> "[CQ:rich,data=${content.escape()}]"
+        is MessageSource -> ""
+        is QuoteReply -> ""
+        is Voice -> "[CQ:voice,url=${url},md5=${md5},file=${fileName}]"
+        else -> "此处消息的转义尚未被插件支持"
+    }
 }
 
 suspend fun codeToChain(bot: Bot, message: String, contact: Contact?): MessageChain {
